% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{get_base_url}
\alias{get_base_url}
\title{Get the base url for a taxon}
\usage{
get_base_url(any = NULL, family = NULL, institution_code = NULL,
  offset = 0)
}
\arguments{
\item{any}{A search query to be matched against all fields.}

\item{family}{Family name.}

\item{institution_code}{An institution code.}

\item{offset}{Page offset}
}
\description{
Get the base url for a taxon
}
