% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{get_image_urls}
\alias{get_image_urls}
\title{Get the urls for a taxon}
\usage{
get_image_urls(any = NULL, family = NULL, institution_code = NULL)
}
\arguments{
\item{any}{A search query to be matched against all fields.}

\item{family}{Family name.}

\item{institution_code}{An institution code.}
}
\description{
Get the urls for a taxon
}
