% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{get_images}
\alias{get_images}
\title{Download images from the Reflora Website.}
\usage{
get_images(any = NULL, family = NULL, institution_code = NULL)
}
\arguments{
\item{any}{A search query to be matched against all fields.}

\item{family}{Family name.}

\item{institution_code}{An institution code.}
}
\value{
\code{get_base_url} Returns the base url for a given taxon. Not very
useful by itself. \code{get_number_of_images} returns the number of images available
for a given taxon. \code{get_image_urls} returns the urls for all images of a given
taxon. \code{get_images} returns NULL and downloads all imagens for the taxon.
}
\description{
Download images from the Reflora Website.
}
\examples{
taxon <- "Miconia albicans"
\dontrun{
get_number_of_images(taxon)
get_image_urls(taxon)
get_images(taxon)
}
}
