% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{get_number_of_images}
\alias{get_number_of_images}
\title{Get the number of images for a given taxon}
\usage{
get_number_of_images(any = NULL, family = NULL, institution_code = NULL)
}
\arguments{
\item{any}{A search query to be matched against all fields.}

\item{family}{Family name.}

\item{institution_code}{An institution code.}
}
\description{
Get the number of images for a given taxon
}
