% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{reflora}
\alias{reflora}
\alias{reflora-package}
\alias{reflora-package}
\title{Package reflora}
\description{
Download specimens images from the Brazilian Reflora Program
(http://www.splink.org.br)
}
\details{
This package contains helper functions to determine the number of images of a
given taxon e download all of them in the highest resolution available. All
images are restricted to specimens that were collected in Brazil (but might have
been deposited in herbaria in other countries).

\itemize{
  \item{get the number of images of a given taxon available}
  \item{download images of a given taxon}
}
}
