% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{get_base_url}
\alias{get_base_url}
\title{Get the base url for a taxon}
\usage{
get_base_url(taxon, offset = 0)
}
\arguments{
\item{taxon}{A unique taxon name.}

\item{offset}{Page offset}
}
\description{
Get the base url for a taxon
}
