% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{get_base_url}
\alias{get_base_url}
\title{Download images from the Reflora Website.}
\usage{
get_base_url(taxon, offset = 0)
}
\arguments{
\item{taxon}{A unique taxon name. Can be anything (species, family, order).}
}
\value{
\code{get_base_url} Returns the base url for a given taxon. Not very
useful by itself. \code{get_number_of_images} returns the number of images available
for a given taxon. \code{get_image_urls} returns the urls for all images of a given
taxon. \code{get_images} returns NULL and downloads all imagens for the taxon.
}
\description{
Download images from the Reflora Website.
}
\examples{
taxon <- "Miconia albicans"
get_base_url(taxon)
\dontrun{
get_number_of_images(taxon)
get_image_urls(taxon)
get_images(taxon)
}
}
