% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_images.R
\name{get_number_of_images}
\alias{get_number_of_images}
\title{Get the number of images for a given taxon}
\usage{
get_number_of_images(taxon)
}
\arguments{
\item{taxon}{A unique taxon name}
}
\description{
Get the number of images for a given taxon
}
